Fixed Bug. Use floats instead of double

+++ RayTracer/src/RayTracer/Classes/hittable.cuh
@@ -7,7 +7,7 @@ struct RT_API hit_record
 {
     point3 p; // 3D point of hit
     vec3 normal; // normal
-    double t; // t value in ray equation
+    float t; // t value in ray equation
     bool front_face;
 
     __device__ inline void set_face_normal(const Ray& r, const vec3& outward_normal) {
@@ -19,5 +19,5 @@ struct RT_API hit_record
 class RT_API Hittable
 {
 public:
-    __device__ virtual bool hit(const Ray& r, double t_min, double t_max, hit_record& rec) const = 0;
+    __device__ virtual bool hit(const Ray& r, float t_min, float t_max, hit_record& rec) const = 0;
 };
