+++ RayTracer/src/RayTracer/Classes/misc.cuh
@@ -9,8 +9,8 @@
 
 // Utility Functions
 
-__device__ inline double degrees_to_radians(double degrees) {
-    return degrees * 3.1415926535897932385 / 180.0;
+__device__ inline float degrees_to_radians(float degrees) {
+    return float(degrees * 3.1415926535897932385 / 180.0);
 }
 
 // Common Headers
